Move statics for `GetStateValidators` to `SharedApiTypes` (#7931)
